test(cli): split Vite plugin integration hotspots
